requirements.txt
flask
scikit-learn
numpy
pytesseract
Pillow
